Several adjustments:
- created .zshrc_secrets loading logic, alongside with a .gitignore
  file that prevents it from being commited to source control.
- Moved some mac specific aliases to the aliases.zsh file, and made them
  only available on mac.
- Removed the requirement for the hub package.

+++ install.sh
@@ -2,7 +2,12 @@
 ln -s -f $HOME/Development/configuration/terminal-config/.antigenrc ~/.antigenrc
 ln -s -f $HOME/Development/configuration/terminal-config/.ackrc ~/.ackrc
 ln -s -f $HOME/Development/configuration/terminal-config/.zshenv ~/.zshenv
-ln -s -f $HOME/Development/configuration/terminal-config/.zshrc ~/.zshrc
+ln -s -f $HOME/Development/configuration/terminal-config/.zshrc_secrets ~/.zshrc_secrets
+
+# Make sure that .zshrc_secrets exists before creating the symlink
+if [ -e "$HOME/.zshrc_secrets" ]; then
+    ln -s -f $HOME/Development/configuration/terminal-config/.zshrc ~/.zshrc
+fi
 echo "Created configuration symlinks"
 
 # Step 2: Create symlinks for global ignore files
@@ -24,4 +29,3 @@ ln -s -f $HOME/Development/configuration/terminal-config/zsh-custom/* ~/zsh-cust
 echo "Added zsh-custom"
 
 echo "Installation completed"
-
